refactor(common/lmlayer): abstracts interface b/t inner and outer layers
