+++ common/predictive-text/worker/index.ts
@@ -349,7 +349,7 @@ if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
   module.exports['ModelCompositor'] = ModelCompositor;
 } else if (typeof self !== 'undefined' && 'postMessage' in self) {
   // Automatically install if we're in a Web Worker.
-  LMLayerWorker.install(self as DedicatedWorkerGlobalScope);
+  LMLayerWorker.install(self as any); // really, 'as typeof globalThis', but we're currently getting TS errors from use of that.
 } else {
   //@ts-ignore
   window.LMLayerWorker = LMLayerWorker;
