+++ common/predictive-text/workerFactory.ts
@@ -1,0 +1,5 @@
+namespace com.keyman.text.prediction {
+  export interface WorkerFactory {
+    constructInstance(): Worker;
+  }
+}
